Fixed bug

+++ MGen/GLibrary/GMidi.cpp
@@ -549,7 +549,8 @@ void CGMidi::LoadCantus(CString path)
 					else {
 						if (nid < MIN_CANTUS_SIZE) {
 							CString st;
-							st.Format("Melody #%d is shorter (%d steps) than minimum length (%d steps): tick %d, track %d, chan %d, tpc %d (mul %.03f) in file %s", cantus.size(), nid, MIN_CANTUS_SIZE, mev->tick, track, mev->getChannel(), tpc, midifile_in_mul, path);
+							st.Format("Melody #%d is shorter (%d steps) than minimum length (%d steps): tick %d, track %d, chan %d, tpc %d (mul %.03f) in file %s", 
+								cantus.size(), nid, (int)MIN_CANTUS_SIZE, mev->tick, track, mev->getChannel(), tpc, midifile_in_mul, path);
 							WriteLog(5, st);
 						}
 					}
@@ -619,7 +620,7 @@ void CGMidi::LoadCantus(CString path)
 		else {
 			if (nid < MIN_CANTUS_SIZE) {
 				CString st;
-				st.Format("Melody #%d is shorter (%d steps) than minimum length (%d steps): tick %d, track %d, tpc %d (mul %.03f) in file %s. Not loaded.", cantus.size(), nid, MIN_CANTUS_SIZE, last_tick, track, tpc, midifile_in_mul, path);
+				st.Format("Melody #%d is shorter (%d steps) than minimum length (%d steps): tick %d, track %d, tpc %d (mul %.03f) in file %s. Not loaded.", cantus.size(), nid, (int)MIN_CANTUS_SIZE, last_tick, track, tpc, midifile_in_mul, path);
 				WriteLog(5, st);
 			}
 		}
@@ -733,7 +734,7 @@ void CGMidi::LoadCP(CString path)
 					else {
 						if (inter.size() < MIN_CP_SIZE) {
 							CString st;
-							st.Format("Counterpoint #%d is shorter (%d steps) than minimum length (%d steps): tick %d, track %d, chan %d, tpc %d (mul %.03f) in file %s", cantus.size(), inter.size(), MIN_CP_SIZE, mev->tick, track, mev->getChannel(), tpc, midifile_in_mul, path);
+							st.Format("Counterpoint #%d is shorter (%d steps) than minimum length (%d steps): tick %d, track %d, chan %d, tpc %d (mul %.03f) in file %s", cantus.size(), inter.size(), (int)MIN_CP_SIZE, mev->tick, track, mev->getChannel(), tpc, midifile_in_mul, path);
 							WriteLog(5, st);
 						}
 					}
@@ -812,7 +813,7 @@ void CGMidi::LoadCP(CString path)
 		else {
 			if (inter.size() < MIN_CP_SIZE) {
 				CString st;
-				st.Format("Counterpoint #%d is shorter (%d steps) than minimum length (%d steps): tick %d, track %d, tpc %d (mul %.03f) in file %s", cantus.size(), inter.size(), MIN_CP_SIZE, last_tick, track, tpc, midifile_in_mul, path);
+				st.Format("Counterpoint #%d is shorter (%d steps) than minimum length (%d steps): tick %d, track %d, tpc %d (mul %.03f) in file %s", cantus.size(), inter.size(), (int)MIN_CP_SIZE, last_tick, track, tpc, midifile_in_mul, path);
 				WriteLog(5, st);
 			}
 		}
